Move/rename interface.
